+++ src/core/v-box.ts
@@ -53,7 +53,7 @@ function getVListChildrenAndDepth(
   params: VBoxParam
 ): [
   children: null | (VBoxChild | VBoxElementAndShift)[] | VBoxChild[],
-  depth: number
+  depth: number,
 ] {
   if ('individualShift' in params) {
     const oldChildren = params.individualShift;
@@ -122,6 +122,7 @@ function makeRows(
   const [children, depth] = getVListChildrenAndDepth(params);
   if (!children) return [[], 0, 0];
 
+  const pstrut = new Box(null, { classes: 'pstrut' });
   // Create a strut that is taller than any list item. The strut is added to
   // each item, where it will determine the item's baseline. Since it has
   // `overflow:hidden`, the strut's top edge will sit on the item's line box's
@@ -137,7 +138,8 @@ function makeRows(
     }
   }
   pstrutSize += 2;
-  const pstrut = new Box(null, { classes: 'pstrut', height: pstrutSize });
+
+  pstrut.height = pstrutSize;
   pstrut.setStyle('height', pstrutSize, 'em');
 
   // Create a new list of actual children at the correct offsets
@@ -150,6 +152,7 @@ function makeRows(
     if (typeof child === 'number') currPos += child;
     else {
       const box = child.box;
+
       const classes = child.classes ?? [];
 
       const childWrap = new Box([pstrut, box], {
@@ -173,14 +176,21 @@ function makeRows(
     maxPos = Math.max(maxPos, currPos);
   }
 
+  realChildren.forEach((child) => {
+    child.softWidth = width;
+  });
+
   // The vlist contents go in a table-cell with `vertical-align:bottom`.
   // This cell's bottom edge will determine the containing table's baseline
   // without overly expanding the containing line-box.
   const vlist = new Box(realChildren, { classes: 'vlist' });
-  vlist.width = width;
+  vlist.softWidth = width;
+  // list.children!.reduce(
+  //   (acc, row) => Math.max(acc, row.width),
+  //   0
+  // );
   vlist.height = maxPos;
   vlist.setStyle('height', maxPos, 'em');
-
   // A second row is used if necessary to represent the vlist's depth.
   if (minPos >= 0)
     return [[new Box(vlist, { classes: 'vlist-r' })], maxPos, -minPos];
@@ -199,9 +209,10 @@ function makeRows(
   const topStrut = new Box(0x200b, {
     classes: 'vlist-s',
     maxFontSize: 0,
-    height: 0,
-    depth: 0,
   });
+  topStrut.softWidth = 0;
+  topStrut.height = 0;
+  topStrut.depth = 0;
 
   return [
     [
@@ -227,9 +238,15 @@ export class VBox extends Box {
         ' vlist-t' +
         (rows.length === 2 ? ' vlist-t2' : ''),
     });
+
     this.height = height;
     this.depth = depth;
-    this.width = rows.reduce((acc, row) => Math.max(acc, row.width), 0);
+    this.softWidth = rows.reduce((acc, row) => Math.max(acc, row.width), 0);
+    // this.width = this.children!.reduce(
+    //   (acc, row) => Math.max(acc, row.width),
+    //   0
+    // );
+    // for (const child of this.children!) child.width = this.width;
   }
 }