+++ src/editor/model.ts
@@ -4,7 +4,6 @@ export { ModelPrivate } from '../editor-model/model-private';
 export * from '../editor-model/array';
 export * from '../editor-model/delete';
 export * from '../editor-model/composition';
-export * from '../editor-model/listeners';
 export * from '../editor-model/selection';
 export * from '../editor-model/styling';