+++ src/vue-mathlive.js
@@ -49,7 +49,6 @@ export default {
     vue.component('mathlive-mathfield', this);
   },
   watch: {
-
     value(newValue) {
       // When the `value` prop (from the model) is modified
       // update the mathfield to stay in sync, but don't send back content